+++ compiler/tests-ocaml/expect.ml
@@ -1,6 +1,6 @@
 [@@@ocaml.alert "-unsafe_multidomain"]
 
-let () = Js_of_ocaml_toplevel_common.JsooTop.initialize ()
+let () = Js_of_ocaml_toplevel_common.Direct.initialize ()
 
 let () = Printexc.register_printer (fun x ->
              match Jsoo_runtime.Error.of_exn x with