+++ toplevel/lib-toplevel/direct.mli
@@ -1,6 +1,6 @@
 (* Js_of_ocaml library
  * http://www.ocsigen.org/js_of_ocaml/
- * Copyright (C) 2017 Hugo Heuzard
+ * Copyright (C) 2014 Hugo Heuzard
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,15 +17,22 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *)
 
-module Direct = Js_of_ocaml_toplevel_common.Direct
-module Wrapped = Js_of_ocaml_toplevel_common.Wrapped
-module Wrapped_intf = Js_of_ocaml_toplevel_common.Wrapped_intf
+open Format
+(** Helper for Js_of_ocaml Toplevel. *)
 
-module JsooTop = Direct
-[@@deprecated "Renamed to Direct; use Js_of_ocaml_toplevel.Direct."]
+val use : formatter -> string -> bool
+(** [use fmt content] Execute commands [content]. It does not print types nor values. *)
 
-module JsooTopWrapped = Wrapped
-[@@deprecated "Renamed to Wrapped; use Js_of_ocaml_toplevel.Wrapped."]
+val execute :
+     bool
+  -> ?pp_code:formatter
+  -> ?highlight_location:(Location.t -> unit)
+  -> formatter
+  -> string
+  -> unit
+(** [execute print fmt content] Execute [content].
+    [print] says whether the values and types of the results should be printed.
+    [pp_code] formatter can be use to output ocaml source during lexing. *)
 
-module JsooTopIntf = Wrapped_intf
-[@@deprecated "Renamed to Wrapped_intf; use Js_of_ocaml_toplevel.Wrapped_intf."]
+val initialize : unit -> unit
+(** Initialize Js_of_ocaml toplevel. Idempotent. *)